refactor: rename tedge_config_cli -> tedge_toml to go with system_toml
