+++ crates/common/tedge_config/src/tedge_toml/error.rs
@@ -13,7 +13,7 @@ pub enum TEdgeConfigError {
     FromConfigSetting(#[from] crate::ConfigSettingError),
 
     #[error(transparent)]
-    FromInvalidConfigUrl(#[from] crate::tedge_config_cli::models::InvalidConnectUrl),
+    FromInvalidConfigUrl(#[from] crate::tedge_toml::models::InvalidConnectUrl),
 
     #[error("Config file not found: {0}")]
     ConfigFileNotFound(std::path::PathBuf),
@@ -31,10 +31,46 @@ pub enum TEdgeConfigError {
     DirNotFound(#[from] tedge_utils::paths::PathsError),
 
     #[error(transparent)]
-    FromParseHostPortError(#[from] crate::tedge_config_cli::models::host_port::ParseHostPortError),
+    FromParseHostPortError(#[from] crate::tedge_toml::models::host_port::ParseHostPortError),
 
     #[error(transparent)]
     FromAtomFileError(#[from] tedge_utils::fs::AtomFileError),
+}
+
+pub type ConfigSettingResult<T> = Result<T, ConfigSettingError>;
+
+#[derive(thiserror::Error, Debug)]
+/// An error encountered while updating a value in tedge.toml
+pub enum ConfigSettingError {
+    #[error(
+        r#"A value for `{key}` is missing.
+    A value can be set with `tedge config set {key} <value>`"#
+    )]
+    ConfigNotSet { key: &'static str },
+
+    #[error("Readonly setting: {message}")]
+    ReadonlySetting { message: &'static str },
+
+    #[error("Conversion from String failed")]
+    ConversionFromStringFailed,
+
+    #[error("Conversion into String failed")]
+    ConversionIntoStringFailed,
+
+    #[error("Derivation for `{key}` failed: {cause}")]
+    DerivationFailed { key: &'static str, cause: String },
+
+    #[error("Config value {key}, cannot be configured: {message} ")]
+    SettingIsNotConfigurable {
+        key: &'static str,
+        message: &'static str,
+    },
+
+    #[error("An error occurred: {msg}")]
+    Other { msg: &'static str },
+
+    #[error(transparent)]
+    Write(#[from] crate::WriteError),
 }
 
 impl TEdgeConfigError {
