+++ crates/common/tedge_config/src/tedge_toml/mod.rs
@@ -1,7 +1,5 @@
-pub mod config_setting;
 pub mod error;
-pub mod tedge_config_location;
-
 mod figment;
 pub mod models;
 pub mod tedge_config;
+pub mod tedge_config_location;
