+++ crates/common/tedge_config/src/tedge_toml/models/c8y_software_management.rs
