+++ crates/core/tedge/src/cli/connect/error.rs
@@ -1,5 +1,4 @@
 use rumqttc::tokio_rustls::rustls;
-use tedge_config::mqtt_config::MqttConfigBuildError;
 
 #[derive(thiserror::Error, Debug)]
 pub enum ConnectError {
@@ -17,9 +16,6 @@ pub enum ConnectError {
 
     #[error(transparent)]
     MqttClient(#[from] rumqttc::ClientError),
-
-    #[error("Can't crate MQTT config")]
-    CreateMqttConfig(#[from] MqttConfigBuildError),
 
     #[error("Can't create TLS config")]
     CreateTlsConfig(#[from] rustls::Error),
